+++ src/main/java/com/texoit/goldenraspberryawardsapi/config/producer/InsertProducerConfig.java
@@ -1,7 +1,7 @@
 package com.texoit.goldenraspberryawardsapi.config.producer;
 
 import com.texoit.goldenraspberryawardsapi.adapters.out.producer.InsertProducerAdapter;
-import com.texoit.goldenraspberryawardsapi.application.core.usecase.movie.InsertProducerUseCase;
+import com.texoit.goldenraspberryawardsapi.application.core.usecase.producer.InsertProducerUseCase;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
